Changing the VM word from uint64_t to uint32_t.

+++ box/box/vm.h
@@ -105,7 +105,7 @@ typedef unsigned char BoxVMByte;
 
 typedef char BoxVMSByte;
 
-typedef uint64_t BoxVMWord;
+typedef uint32_t BoxVMWord;
 
 
 #  define BoxVMWord_Fmt "%8.8lx"
